requirements.txt
discord.py
nltk